+++ lib/screens_pages/sign_up_page.dart
@@ -5,7 +5,7 @@ import 'package:google_fonts/google_fonts.dart';   //Fonts
 import 'package:simple_chat/standalone_methods_functions/general_methods.dart';
 import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
 import 'package:simple_chat/configuration/config_invoke.dart';
-import 'package:simple_chat/utils/alert_dialog_list.dart';
+import 'package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart';
 import 'package:simple_chat/utils/logger_config.dart';
 import 'package:simple_chat/screens_pages/log_in_page.dart';
